update auth routes
