+++ frontend/src/App.tsx
@@ -5,10 +5,10 @@ import { AuthProvider } from "./hooks/useAuth";
 import { ChatSocketProvider } from "./hooks/useChatSocket";
 import { FriendshipSocketProvider } from "./hooks/useFriendshipSocket";
 import { GameSocketProvider } from "./hooks/useGameSocket";
-import Login from "./pages/auth/Login";
-import Register from "./pages/auth/Register";
-import Setup from "./pages/auth/Setup";
-import TwoFaAuth from "./pages/auth/TwoFaAuth";
+import Login from "./pages/auth/login/Login";
+import TwoFaAuth from "./pages/auth/login/TwoFaAuth";
+import Register from "./pages/auth/register/Register";
+import Setup from "./pages/auth/register/Setup";
 import Chat from "./pages/chat/Chat";
 import PrivateChat from "./pages/chat/PrivateChat";
 import Game from "./pages/game/Game";
