+++ frontend/src/components/layouts/layout/Layout.tsx
@@ -40,13 +40,13 @@ function Layout() {
           <Header />
           <main className="flex-1 p-4 md:p-6">{isHomePage() ? <Home /> : <Outlet />}</main>
           <Footer />
-          <Notify />
         </div>
       ) : (
         <main className="min-h-screen flex flex-col justify-center items-center">
           <Outlet />
         </main>
       )}
+      <Notify />
     </>
   );
 }
