--- frontend/src/pages/auth/Auth.module.scss
@@ -1,207 +1,0 @@
-@import "../../styles/config.scss";
-
-.container {
-  height: calc(100vh - 2rem);
-  flex-direction: column;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 0.5rem;
-
-  .form {
-    max-width: 22rem;
-    padding: 2rem;
-    flex-direction: column;
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-    border-radius: $rounded-2xl;
-    background: $bg-secondary-color;
-    color: $text-primary-color;
-    text-align: center;
-
-    .error {
-      width: 22rem;
-      padding: 0.5rem 0.5rem;
-      box-sizing: border-box;
-      border: none;
-      border-radius: $rounded-md;
-      background: #f8d7da;
-
-      p {
-        font-size: map-get($text-sm, font-size);
-        line-height: map-get($text-sm, line-height);
-        font-weight: $font-normal;
-        text-align: center;
-        color: #721c24;
-      }
-    }
-
-    .input {
-      flex-direction: column;
-      display: flex;
-      align-items: left;
-      gap: 0.25rem;
-
-      label {
-        display: block;
-        font-size: map-get($text-sm, font-size);
-        line-height: map-get($text-sm, line-height);
-        font-weight: $font-medium;
-        text-align: left;
-      }
-
-      input {
-        width: 22rem;
-        height: 2.5rem;
-        padding: 0 1rem;
-        box-sizing: border-box;
-        border: none;
-        border-radius: $rounded-md;
-        outline: none;
-        background: $extra-tertiary-color;
-        font-size: map-get($text-sm, font-size);
-        line-height: map-get($text-sm, line-height);
-        font-weight: $font-normal;
-        color: $text-primary-color;
-      }
-    }
-
-    .buttons {
-      margin-top: 0.75rem;
-      flex-direction: column;
-      display: flex;
-      align-items: center;
-      gap: 0.65rem;
-
-      button {
-        width: 22rem;
-        height: 2.5rem;
-        box-sizing: border-box;
-        border: none;
-        border-radius: $rounded-md;
-        transition: background 0.3s ease;
-        font-size: map-get($text-base, font-size);
-        line-height: map-get($text-base, line-height);
-        font-weight: $font-medium;
-        color: white;
-        cursor: pointer;
-
-        &.submitBtn {
-          background: $extra-secondary-color;
-
-          &:hover {
-            background: $extra-primary-color;
-          }
-        }
-
-        &.fortyTwoBtn {
-          background: #36babb;
-
-          &:hover {
-            background: #48c7c7;
-          }
-        }
-      }
-    }
-
-    .help {
-      span {
-        font-size: map-get($text-sm, font-size);
-        line-height: map-get($text-sm, line-height);
-        font-weight: $font-light;
-        color: #9ca3af;
-      }
-
-      .link {
-        margin-left: 0.3rem;
-        font-size: map-get($text-sm, font-size);
-        line-height: map-get($text-sm, line-height);
-        font-weight: $font-medium;
-        text-decoration: none;
-        color: #9ca3af;
-      }
-    }
-  }
-}
-
-@media (max-width: 768px) {
-  .container {
-    .form {
-      padding: 1rem;
-      gap: 0.5rem;
-
-      h1 {
-        margin: 0;
-        font-size: map-get($text-lg, font-size);
-        line-height: map-get($text-lg, line-height);
-        font-weight: $font-semibold;
-      }
-
-      p {
-        margin: 0;
-        font-size: map-get($text-xs, font-size);
-        line-height: map-get($text-xs, line-height);
-        font-weight: $font-normal;
-      }
-
-      .error {
-        width: 15rem;
-
-        p {
-          font-size: map-get($text-xs, font-size);
-          line-height: map-get($text-xs, line-height);
-          font-weight: $font-normal;
-        }
-      }
-
-      .input {
-        label {
-          display: block;
-          font-size: map-get($text-xs, font-size);
-          line-height: map-get($text-xs, line-height);
-          font-weight: $font-medium;
-          text-align: left;
-        }
-
-        input {
-          width: 15rem;
-          height: 2.25rem;
-          font-size: map-get($text-xs, font-size);
-          line-height: map-get($text-xs, line-height);
-          font-weight: $font-normal;
-        }
-      }
-
-      .buttons {
-        margin-top: 0.5rem;
-
-        button {
-          width: 15rem;
-          height: 2.25rem;
-          font-size: map-get($text-sm, font-size);
-          line-height: map-get($text-sm, line-height);
-          font-weight: $font-medium;
-        }
-      }
-
-      .help {
-        span {
-          font-size: map-get($text-xs, font-size);
-          line-height: map-get($text-xs, line-height);
-          font-weight: $font-light;
-          color: #9ca3af;
-        }
-
-        .link {
-          margin-left: 0.3rem;
-          font-size: map-get($text-xs, font-size);
-          line-height: map-get($text-xs, line-height);
-          font-weight: $font-medium;
-          text-decoration: none;
-          color: #9ca3af;
-        }
-      }
-    }
-  }
-}
