+++ frontend/src/pages/settings/TwoFaSetupDialog.tsx
@@ -95,7 +95,6 @@ export const TwoFaSetupDialog = () => {
                 </InputOTPGroup>
               </InputOTP>
             </div>
-
             <div className="ml-auto flex items-center gap-2">
               <DialogClose asChild>
                 <Button type="button" variant="ghost">
